feat: Fetch and store project commits
* The commits table's primary key is the commit's `short_id`.
* Author's name is used and tied to the member's name. Note that some 
users has different names in their git config rather than their 
account's name on Gitlab. At this point I am not sure what's the side 
effect of that or how shall I handle it.
* Authored date is stored because we can use it to request commits after 
that date from Gitlab API
* The request to get new commits will first check if there're any 
existing commits related to that project or not. If there're any commits 
it will use the date of the last commit to request commits made after 
that date only.
* Starting from this point I will clean all the data before storing it 
to keep the DB as small as possible